started implementing DI-system
