+++ src/kentaasvang.Rssh/Interfaces/IAddHandler.cs
@@ -1,0 +1,6 @@
+namespace kentaasvang.Rssh.Interfaces;
+
+public interface IAddHandler
+{
+    public void Handler(string connectionName);
+}
