+++ src/kentaasvang.Rssh/Interfaces/IInputProvider.cs
@@ -1,0 +1,5 @@
+namespace kentaasvang.Rssh.Interfaces;
+public interface IInputProvider
+{
+    public string GetInput();
+}
